core/swapper: moved to core/wallet/swapper
