+++ core/wallet/swapper/immediate/immediate.go
